Rename test files to match pattern "test*.py"
This is needed for automatic discover of tests using
`python -m unittest discover`
